Rename run_aic to RunAIC and move MakePlot from aic_nd.py to utils_nd.py
